Added the ability to play sounds.
Instead of promping for a folder every time we start, I've hardcoded tilt/x.mp3 into the program.
If you want to play these sounds, copy the tilt/ that I've committed to the root
of your /sdcard/ on your phone (It should be alongside Downloads and Music)

Android seems to dislike the we I'm playing audio, and after ~20 or 40 plays it goes
mute, and debugging can't really explain why. So if it goes mute for you kill
the app and then reopen it. (Settings->Apps->Tilt->Force Quit)

Also, the filenames are the directions that play them, so if you dislike the
included sounds, feel free to replace them. Just keep in mind case matters.